Use gaol as the sandboxing mechanism
This commit signals the intent of this project to explore `gaol` as the
mechanism by which builds are sandboxed.

The README.md for `gaol` carries the following warning:

> gaol is only lightly reviewed for correctness and security.
> It should not be considered mature or "battle-tested".
> Use at your own risk.

However, cursory review of gaol's source code shows it's a relatively
simple cross-platform sandboxing solution which provides pure Rust
wrappers for the underlying OS facilities sandboxing solutions should be
leveraging.

As `gaol` is used by Servo, it's something the Rust team is probably
already familiar with, which increases the chances of upstreaming a
sandboxing solution into Cargo proper, a potential goal for this
project.